backend/requirements.txt
fastapi
uvicorn
openai
elevenlabs
supabase
python-dotenv
requests
pypdf2
